pyproject.toml
[tool.poetry]
name = "digitalartapi"
version = "0.1.0"
description = ""
authors = ["PatkaSm <patrycja.s449@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.5"
djangorestframework = "^3.11.0"
Pillow = "^7.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
